added cleanupacme.py
this script will reference subdomainlist.json, and the acme.json generated by traefik.
it will removed any entries in acme.json that do not match with subdomainlist.json.
this will help ensure traefik does not keep requesting certs from letsencrypt that no longer have a DNS record.

this will need to be more elegantly incorporated into the docker build and ddns app at a future time -- for now it is manually run, and files need to be manually moved around.